renamed module archives to farchives to prevent confusion wich archive plugin
